% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bivalvia.R
\name{bivalvia}
\alias{bivalvia}
\title{Bivalvia Key}
\usage{
bivalvia(image.folder)
}
\arguments{
\item{image.folder}{Path to the folder that contains your identification images}
}
\value{
A list of your key choices as well as a species Identification!
}
\description{
This function allows you to identify your bivalve based on their morphology. You should start with an empty
folder titled "Bivalvia" where you will eventually store photos of the morphological structures you've used
to identify your bivalves. KEEP SEPARATE FROM YOUR OTHER FOLDERS/PICUTRES OF ORGANISMS! This function
requires a single input: the path to this folder. It's ok if it's empty to begin with, but in order to run the
function you do need to provide it's path on your computer. This path should be contained in quotes and end in
a forward slash "/". Once you provide the path, you can begin keying your bivalve! As you move through the
identification, you will need to enter numbers based on the interactive questions. Once you populate images
into your folder, you will see them plot side by side on the questions they apply to! Important: when naming
images, use the following structure "choice_#.jpg" where the pound sign "#" should be replaced by the number
you would choose in the key based on the morphological structure you've photographed. You should have no more
than one photo per number.
}
